Move compiletest CLI parsing to `cli.rs`
