+++ src/tools/compiletest/src/directives/tests.rs
@@ -267,7 +267,7 @@ impl ConfigBuilder {
         args.push("--rustc-path".to_string());
         args.push(std::env::var("TEST_RUSTC").expect("must be configured by bootstrap"));
 
-        crate::parse_config(args)
+        crate::cli::parse_config(args)
     }
 }
